Rename `cltool` to `command_line_tool`
* to avoid confusion with cwltool itself
